Wrap login fragment layout in ScrollView and set windowSoftInputMode for activities
The login fragment layout is now wrapped in a `ScrollView` to allow scrolling when the keyboard is visible.
The `android:windowSoftInputMode` attribute has been set to `adjustPan` for `HomeActivity`, `authActivity`, `AdminActivity`, `AddAdminActivity`, and `MainActivity` in the `AndroidManifest.xml`. This ensures that the layout pans to make room for the soft keyboard without resizing.